Moving rs2_project_point_to_pixel and rs2_deproject_pixel_to_point to utils.cpp
